% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopBaseCorrect.R
\name{dopBaseCorrect}
\alias{dopBaseCorrect}
\title{Baseline correct}
\usage{
dopBaseCorrect(dop, baseline)
}
\arguments{
\item{dop}{list DOPOSSCI list of settings + data etc.}

\item{baseline}{numeric Lower and Upper limits of baseline period in seconds}
}
\value{
dop
}
\description{
Subtract the average of a baseline period from each waveform.
"dop" variable is returned with dop$epoch$base which is the baseline-corrected data.
}
