% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopEpoch.R
\name{dopEpoch}
\alias{dopEpoch}
\title{Epoch continuous fTCD data}
\usage{
dopEpoch(dop)
}
\arguments{
\item{dop}{list DOPOSSCI list of settings + data etc.}
}
\value{
dop
}
\description{
Examines the event channel to create a matrix of sample by epoch sized matrix
}
