% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopHeartCycle.R
\name{dopHeartCycle}
\alias{dopHeartCycle}
\title{Correct for the Heart Cycle/Heart Best in fTCD signal}
\usage{
dopHeartCycle(dop, heart_cycle_window = 0.26, channel_name = "velocity")
}
\arguments{
\item{dop}{list}

\item{heart_cycle_window}{numeric}

\item{channel_name}{character}
}
\value{
dop list
}
\description{
The raw fTCD signal include heart beats which are best smoothed over for most fTCD processing
}
\details{
The idea comes from Deppe et al. (1997)

References
Deppe, M., Knecht, S., Henningsen, H., & Ringelstein, E. B. (1997). AVERAGE: A Windows® program for automated analysis of event related cerebral blood flow. Journal of Neuroscience Methods, 75(2), 147–154. https://doi.org/10.1016/S0165-0270(97)00067-8
}
