% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopRead.R
\name{dopRead}
\alias{dopRead}
\title{Import functional transcranial Doppler ultrasound (fTCD) data}
\usage{
dopRead(file_name, fTCD_system)
}
\arguments{
\item{file_name}{character File name or path/file name to fTCD file}

\item{fTCD_system}{character The fTCD recording system/hardware. Options: "DWL", "Tari"}
}
\value{
df = dataframe of imported data
}
\description{
dopRead acts as a switch to pass a file name to an fTCD-system specific import function
}
\examples{

}
