% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopReadTari.R
\name{dopReadTari}
\alias{dopReadTari}
\title{Import functional transcranial Doppler ultrasound (fTCD) data: Tari system}
\usage{
dopReadTari(file_name)
}
\arguments{
\item{file_name}{character File name or path/file name to fTCD file}
}
\value{
df dataframe Data frame on imported data
}
\description{
Import functional transcranial Doppler ultrasound (fTCD) data: Tari system
}
