% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopTrim.R
\name{dopTrim}
\alias{dopTrim}
\title{Trim the ends of the fTCD data}
\usage{
dopTrim(dop)
}
\arguments{
\item{dop}{list DOPOSCCI list/stucture holding settings}
}
\value{
dop list Updated DOPOSCCI list
}
\description{
Trim the ends of the fTCD data
}
